% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFootprints.R
\name{getFootprints}
\alias{getFootprints}
\title{Get Footprints}
\usage{
getFootprints(files, method = max)
}
\arguments{
\item{files}{Character. Vector of input csv filenames (from EarthExplorer)}

\item{method}{Function. Function to use to calculate vertices (default is median)}
}
\description{
Get individual scene footprints from several EarthExplorer csv metadata files
}
\examples{
fl <- list.files('inst/extdata/', pattern = glob2rx('*.csv'), full.names = TRUE)
fp <- getFootprints(fl)
plot(fp)

# or if you want to get fancy....
library(mapview)
mapView(fp)
}

