% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsedata.R
\name{parsedata}
\alias{parsedata}
\title{Parse metadata}
\usage{
parsedata(files)
}
\arguments{
\item{files}{Character. Vector of input csv filenames (from EarthExplorer)}
}
\value{
\code{data.frame} with cloud cover per date per image
}
\description{
Parse select attributes from metadata csv files
}
\examples{
fl <- list.files('inst/extdata/', pattern = glob2rx('*.csv'), full.names = TRUE)
df <- parsedata(fl)
head(df)
tail(df)

# plot cloud cover per DOY per year per path/row
library(magrittr)
library(ggplot2)
df$year <- format(df$date, format = "\%Y") \%>\% as.numeric()
df$doy <- format(df$date, format = "\%j") \%>\% as.numeric()
df$pr <- sprintf("\%s-\%s", df$path, df$row) \%>\% factor()
p <- ggplot(data = df, aes(x = doy, y = year)) +
   geom_point(aes(size = cloud_cover, colour = sensor), alpha = 0.5) +
   facet_wrap(~ pr, nc = 1)
p
}

