% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsedata.R
\name{parsedata}
\alias{parsedata}
\title{Parse metadata}
\usage{
parsedata(files)
}
\arguments{
\item{files}{Character. Vector of input csv filenames (from EarthExplorer)}
}
\value{
\code{data.frame} with cloud cover per date per image
}
\description{
Parse select attributes from metadata csv files
}
\examples{
fl <- list.files('inst/extdata/', pattern = glob2rx('*.csv'), full.names = TRUE)
df <- parsedata(fl)
head(df)
tail(df)


}

